+++ mmtbx/model/model.py
@@ -973,7 +973,8 @@ class manager(object):
       result = True
     return result
 
-  def macromolecule_plus_hetatms_by_chain_selections(self, radius=3):
+  def macromolecule_plus_hetatms_by_chain_selections(self, radius=3,
+       max_radius=5):
     """
     Split model into a list of selections where each selection is a
     macromolecule chain plus nearest small molecules (water, ligands, etc).
@@ -985,9 +986,9 @@ class manager(object):
     h = self.get_hierarchy()
     asc = h.atom_selection_cache()
     het_sel = asc.selection("not (protein or nucleotide)")
-    pro_sel = ~het_sel
     size = h.atoms_size()
     get_class = iotbx.pdb.common_residue_names_get_class
+    pro_sel = ~het_sel
     atoms = h.atoms()
     sites_cart = atoms.extract_xyz()
     # list protein chain selections
@@ -999,15 +1000,15 @@ class manager(object):
       psels.append(sel)
     #
     for rg in h.residue_groups():
-      tmp = []
-      for resname in rg.unique_resnames():
-        e1 = get_class(resname) == "common_amino_acid"
-        e2 = get_class(resname) == "common_rna_dna"
-        tmp.extend([e1,e2])
-      if(True in tmp): continue
       residue_b_selection = flex.bool(size, rg.atoms().extract_i_seq())
       around = 0
       radius_ = radius
+      skip = False
+      for i in rg.atoms().extract_i_seq():
+        if(not het_sel[i]):
+          skip = True
+          break
+      if(skip): continue
       while around<1:
         selection_around_residue = sps.pair_generator(
           sites_cart      = sites_cart,
@@ -1019,11 +1020,15 @@ class manager(object):
         selection_around_residue_i = selection_around_residue.iselection()
         around = selection_around_residue_i.size()
         radius_ += 0.5
+        if(radius_>=max_radius): break
       # find which protein chain this hetatm belongs to and add it
-      for psel in psels:
-        if(psel[selection_around_residue_i[0]]):
-          psel = psel.set_selected(residue_b_selection, True)
-          break
+      if(around > 0):
+        for psel in psels:
+          if(psel[selection_around_residue_i[0]]):
+            psel = psel.set_selected(residue_b_selection, True)
+            break
+      else: # sigle atom in space
+        psels.append(residue_b_selection) # just add it as is (alone)
     # checksum
     overlap_sel = psels[0].as_int()
     cntr = 0
@@ -1958,11 +1963,12 @@ class manager(object):
     g = self.get_ncs_groups()
     return g is not None and len(g)>0
 
-  def search_for_ncs(self, params=None, show_groups=False):
+  def search_for_ncs(self, params=None, show_groups=False, ncs_phil_groups=None):
     self._ncs_obj = iotbx.ncs.input(
-      hierarchy = self.get_hierarchy(),
-      params    = params,
-      log       = self.log)
+      hierarchy       = self.get_hierarchy(),
+      ncs_phil_groups = ncs_phil_groups,
+      params          = params,
+      log             = self.log)
     if(self._ncs_obj is not None):
       self._ncs_groups = self.get_ncs_obj().get_ncs_restraints_group_list()
     self._update_master_sel()