+++ mmtbx/reduce/Optimizers.py
@@ -130,7 +130,7 @@ def GetAtomsForConformer(model, conf):
 
 class _SingletonOptimizer(object):
 
-  def __init__(self, model, modelIndex = 0, altID = None,
+  def __init__(self, addFlipMovers, model, modelIndex = 0, altID = None,
                 bondedNeighborDepth = 3,
                 probeRadius = 0.25,
                 useNeutronDistances = False,
@@ -144,6 +144,7 @@ class _SingletonOptimizer(object):
     ignoring their impact on one another.  This means that it will not find the global
     minimum for any multi-Mover Cliques.  Derived classes should override the
     _optimizeCliqueCoarse() and other methods as needed to improve correctness and speed.
+    :param addFlipMovers: Do we add flip Movers along with other types?
     :param model: iotbx model (a group of hierarchy models).  Can be obtained using
     iotbx.map_model_manager.map_model_manager.model().  The model must have Hydrogens,
     which can be added using mmtbx.hydrogens.reduce_hydrogen.place_hydrogens().get_model().
@@ -218,7 +219,7 @@ class _SingletonOptimizer(object):
 
       # Get the list of alternate conformation names present in all chains for this model.
       # If there is more than one result, remove the empty results and then sort them
-      # in reverse order.
+      # in reverse order so we finalize all non-alternate ones to match the first.
       alts = AlternatesInModel(myModel)
       if len(alts) > 1:
         alts.discard("")
@@ -231,7 +232,6 @@ class _SingletonOptimizer(object):
         alts = [altID]
 
       for ai in alts:
-
         # Tell about the run we are currently doing.
         self._infoString += _VerboseCheck(1,"Running Reduce optimization on model index "+str(mi)+
           ", alternate '"+ai+"'\n")
@@ -263,10 +263,15 @@ class _SingletonOptimizer(object):
 
         ################################################################################
         # Get the probeExt.ExtraAtomInfo needed to determine which atoms are potential acceptors.
-        ret = Helpers.getExtraAtomInfo(model)
+        ret = Helpers.getExtraAtomInfo(model) # @todo Pass probePhil here
         self._extraAtomInfo = ret.extraAtomInfo
         self._infoString += ret.warnings
         self._infoString += _ReportTiming("get extra atom info")
+
+        ################################################################################
+        # Initialize any per-alternate data structures now that we have the atoms and
+        # other data structures initialized.
+        self._initializeAlternate()
 
         ################################################################################
         # Find the radius of the largest atom we'll have to deal with.
@@ -279,7 +284,7 @@ class _SingletonOptimizer(object):
         # Get the list of Movers using the _PlaceMovers private function.
         ret = _PlaceMovers(self._atoms, model.rotatable_hd_selection(iselection=True),
                            bondedNeighborLists, self._spatialQuery, self._extraAtomInfo,
-                           self._maximumVDWRadius)
+                           self._maximumVDWRadius, addFlipMovers)
         self._infoString += ret.infoString
         self._movers = ret.moverList
         self._infoString += _VerboseCheck(1,"Inserted "+str(len(self._movers))+" Movers\n")
@@ -546,9 +551,19 @@ class _SingletonOptimizer(object):
     self._atomDump = ""
     return ret
 
+  def _initializeAlternate(self):
+    """
+    Override this method in derived classes.
+    This is a place for derived classes to perform any operations that should be done at
+    the start of every alternate.  For example, initializing per-atom caches.
+    """
+    return
+
   def _setMoverState(self, positionReturn, index):
-    # Move the atoms to their new positions, updating the spatial query structure
-    # by removing the old and adding the new location.
+    """
+      Move the atoms to their new positions, updating the spatial query structure
+      by removing the old and adding the new location.
+    """
     for i, a in enumerate(positionReturn.atoms):
       self._spatialQuery.remove(a)
       a.xyz = positionReturn.positions[index][i]
@@ -685,7 +700,7 @@ class _SingletonOptimizer(object):
     return maxScore
 
 class _BruteForceOptimizer(_SingletonOptimizer):
-  def __init__(self, model, modelIndex = 0, altID = None,
+  def __init__(self, addFlipMovers, model, modelIndex = 0, altID = None,
                 bondedNeighborDepth = 3,
                 probeRadius = 0.25,
                 useNeutronDistances = False,
@@ -696,6 +711,7 @@ class _BruteForceOptimizer(_SingletonOptimizer):
     """Constructor for _BruteForceOptimizer.  This tries all combinations of Mover positions
     within a Clique.  It will be too slow for many files, but it provides a baseline against
     which to compare the results from faster optimizers.
+    :param addFlipMovers: Do we add flip Movers along with other types?
     :param model: iotbx model (a group of hierarchy models).  Can be obtained using
     iotbx.map_model_manager.map_model_manager.model().  The model must have Hydrogens,
     which can be added using mmtbx.hydrogens.reduce_hydrogen.place_hydrogens().get_model().
@@ -724,7 +740,7 @@ class _BruteForceOptimizer(_SingletonOptimizer):
     :param preferenceMagnitude: Multiplier for the preference energies expressed
     by some Movers for particular orientations.
     """
-    super(_BruteForceOptimizer, self).__init__(model, modelIndex = modelIndex, altID = altID,
+    super(_BruteForceOptimizer, self).__init__(addFlipMovers, model, modelIndex = modelIndex, altID = altID,
                 bondedNeighborDepth = bondedNeighborDepth,
                 probeRadius = probeRadius, useNeutronDistances = useNeutronDistances, probeDensity = probeDensity,
                 minOccupancy = minOccupancy, preferenceMagnitude = preferenceMagnitude)
@@ -791,7 +807,7 @@ class _BruteForceOptimizer(_SingletonOptimizer):
     return ret
 
 class _CliqueOptimizer(_BruteForceOptimizer):
-  def __init__(self, model, modelIndex = 0, altID = None,
+  def __init__(self, addFlipMovers, model, modelIndex = 0, altID = None,
                 bondedNeighborDepth = 3,
                 probeRadius = 0.25,
                 useNeutronDistances = False,
@@ -807,6 +823,7 @@ class _CliqueOptimizer(_BruteForceOptimizer):
     and then (with each of the subcomponents in its optimal state) computes the score for the Movers in
     the vertex cut.  Recursion terminates when there are two or fewer Movers in the Clique or when no
     vertex cut can be found; the parent-class Clique solver is used in these cases.
+    :param addFlipMovers: Do we add flip Movers along with other types?
     :param model: iotbx model (a group of hierarchy models).  Can be obtained using
     iotbx.map_model_manager.map_model_manager.model().  The model must have Hydrogens,
     which can be added using mmtbx.hydrogens.reduce_hydrogen.place_hydrogens().get_model().
@@ -835,7 +852,7 @@ class _CliqueOptimizer(_BruteForceOptimizer):
     :param preferenceMagnitude: Multiplier for the preference energies expressed
     by some Movers for particular orientations.
     """
-    super(_CliqueOptimizer, self).__init__(model, modelIndex = modelIndex, altID = altID,
+    super(_CliqueOptimizer, self).__init__(addFlipMovers, model, modelIndex = modelIndex, altID = altID,
                 bondedNeighborDepth = bondedNeighborDepth,
                 probeRadius = probeRadius, useNeutronDistances = useNeutronDistances, probeDensity = probeDensity,
                 minOccupancy = minOccupancy, preferenceMagnitude = preferenceMagnitude)
@@ -934,7 +951,7 @@ class _CliqueOptimizer(_BruteForceOptimizer):
     return ret
 
 class FastOptimizer(_CliqueOptimizer):
-  def __init__(self, model, modelIndex = 0, altID = None,
+  def __init__(self, addFlipMovers, model, modelIndex = 0, altID = None,
                 bondedNeighborDepth = 3,
                 probeRadius = 0.25,
                 useNeutronDistances = False,
@@ -948,6 +965,7 @@ class FastOptimizer(_CliqueOptimizer):
     is overridden to use this cached value in clique optimization (but not in singleton
     or fine optimization) when it has already been computed for a given configuration
     of Movers.
+    :param addFlipMovers: Do we add flip Movers along with other types?
     :param model: iotbx model (a group of hierarchy models).  Can be obtained using
     iotbx.map_model_manager.map_model_manager.model().  The model must have Hydrogens,
     which can be added using mmtbx.hydrogens.reduce_hydrogen.place_hydrogens().get_model().
@@ -983,10 +1001,23 @@ class FastOptimizer(_CliqueOptimizer):
     # it calls.
     self._doScoreCaching = False
 
-    super(FastOptimizer, self).__init__(model, modelIndex = modelIndex, altID = altID,
+    super(FastOptimizer, self).__init__(addFlipMovers, model, modelIndex = modelIndex, altID = altID,
                 bondedNeighborDepth = bondedNeighborDepth,
                 probeRadius = probeRadius, useNeutronDistances = useNeutronDistances, probeDensity = probeDensity,
                 minOccupancy = minOccupancy, preferenceMagnitude = preferenceMagnitude)
+
+  def _initializeAlternate(self):
+    # Ensure that we have a per-atom _scoreCache dictionary that will store already-computed
+    # results for a given atom based on the configurations of the Movers that can affect its
+    # results.  The entries will be empty to start with and will be filled in as they are computed.
+    # We build entries for all atoms, even those not in the Movers to avoid having to traverse
+    # the Movers.
+    # This structure is a dictionary (looked up by atom) of dictionaries (looked up by tuple)
+    # of values (scores).
+    self._scoreCache = {}
+    for a in self._atoms:
+      self._scoreCache[a] = {}
+    return
 
   def _scoreAtom(self, atom):
 
@@ -1007,7 +1038,7 @@ class FastOptimizer(_CliqueOptimizer):
   def _optimizeCliqueCoarse(self, clique):
     """
     The FastOptimizer class generates a per-atom score cache object and uses it along
-    with an overridden _doScoreCaching() method to avoid recomputing scores for atoms where
+    with an overridden _scoreAtom() method to avoid recomputing scores for atoms where
     there has been no change in any of the Movers they depend on.
     It wraps the parent-class method after setting things up to use the cache, and
     then turns off the cache before returning.
@@ -1019,18 +1050,6 @@ class FastOptimizer(_CliqueOptimizer):
     :side_effect: self._highScores is set to the individual score for each of the Movers.
     """
     self._infoString += _VerboseCheck(3,"Optimizing clique of size {} using atom-score cache\n".format(len(list(clique.vertices()))))
-
-    # Ensure that we have a per-atom _scoreCache dictionary that will store already-computed
-    # results for a given atom based on the configurations of the Movers that can affect its
-    # results.  The entries will be empty to start with and will be filled in as they are computed.
-    # We build entries for all atoms, even those not in the Movers to avoid having to traverse
-    # the Movers.
-    # This structure is a dictionary (looked up by atom) of dictionaries (looked up by tuple)
-    # of values (scores).
-    if not hasattr(self, "_scoreCache"):
-      self._scoreCache = {}
-      for a in self._atoms:
-        self._scoreCache[a] = {}
 
     # Call the parent-class optimizer, turning on and off the cache behavior before
     # and after.
@@ -1054,7 +1073,7 @@ class _PlaceMoversReturn(object):
     self.deleteAtoms = deleteAtoms
 
 def _PlaceMovers(atoms, rotatableHydrogenIDs, bondedNeighborLists, spatialQuery, extraAtomInfo,
-                  maxVDWRadius):
+                  maxVDWRadius, addFlipMovers):
   """Produce a list of Movers for atoms in a pdb.hierarchy.conformer that has added Hydrogens.
   :param atoms: flex array of atoms to search.  This must have all Hydrogens needed by the
   Movers present in the structure already.
@@ -1068,6 +1087,7 @@ def _PlaceMovers(atoms, rotatableHydrogenIDs, bondedNeighborLists, spatialQuery,
   information about atoms beyond what is in the pdb.hierarchy.  Used here to determine
   which atoms may be acceptors.
   :param maxVDWRadius: Maximum VdW radius of the atoms.
+  :param addFlipMovers: Do we add flip Movers along with other types?
   :return: _PlaceMoversReturn giving the list of Movers found in the conformation and
   an error string that is empty if no errors are found during the process and which
   has a printable message in case one or more errors are found.
@@ -1199,7 +1219,7 @@ def _PlaceMovers(atoms, rotatableHydrogenIDs, bondedNeighborLists, spatialQuery,
 
     # See if we should insert a MoverNH2Flip here.
     # @todo Is there a more general way than looking for specific names?
-    if (aName == 'ND2' and resName == 'ASN') or (aName == 'NE2' and resName == 'GLN'):
+    if addFlipMovers and ((aName == 'ND2' and resName == 'ASN') or (aName == 'NE2' and resName == 'GLN')):
       try:
         movers.append(Movers.MoverNH2Flip(a, "CA", bondedNeighborLists))
         infoString += _VerboseCheck(1,"Added MoverNH2Flip "+str(len(movers))+" to "+resNameAndID+"\n")
@@ -1294,7 +1314,7 @@ def _PlaceMovers(atoms, rotatableHydrogenIDs, bondedNeighborLists, spatialQuery,
           _modifyIfNeeded(fixUp.atoms[4], coarsePositions[4], fixUp.atoms[5], infoString)
 
           infoString += _VerboseCheck(1,"Set MoverHistidineFlip on "+resNameAndID+" to state "+str(bondedConfig)+"\n")
-        else:
+        elif addFlipMovers: # Add a Histidine flip Mover if we're adding flip Movers
           movers.append(hist)
           infoString += _VerboseCheck(1,"Added MoverHistidineFlip "+str(len(movers))+" to "+resNameAndID+"\n")
       except Exception as e:
@@ -1504,7 +1524,7 @@ END
   atoms = GetAtomsForConformer(firstModel, "")
 
   # Get the probeExt.ExtraAtomInfo needed to determine which atoms are potential acceptors.
-  ret = Helpers.getExtraAtomInfo(model)
+  ret = Helpers.getExtraAtomInfo(model) # @todo Pass probePhil here
   extra = ret.extraAtomInfo
 
   # Get the Cartesian positions of all of the atoms we're considering for this alternate
@@ -1536,7 +1556,7 @@ END
   # Place the movers, which should include only an NH3 rotator because the Histidine flip
   # will be constrained by the ionic bonds.
   ret = _PlaceMovers(atoms, model.rotatable_hd_selection(iselection=True),
-                     bondedNeighborLists, sq, extra, maxVDWRad)
+                     bondedNeighborLists, sq, extra, maxVDWRad, True)
   movers = ret.moverList
   if len(movers) != 1:
     return "Optimizers.Test(): Incorrect number of Movers for 1xso Histidine test"
@@ -1547,10 +1567,10 @@ END
     name = a.name.strip()
     if name in ["ND1", "NE2"]:
       if not extra.getMappingFor(a).isAcceptor:
-        return "Optimizers.Test(): '+a+' in 1xso Histidine test was not an acceptor"
+        return 'Optimizers.Test(): '+name+' in 1xso Histidine test was not an acceptor'
     if name in ["HD1", "HE2"]:
       if not a in ret.deleteAtoms:
-        return "Optimizers.Test(): '+a+' in 1xso Histidine test was not set for deletion"
+        return 'Optimizers.Test(): '+name+' in 1xso Histidine test was not set for deletion'
 
   #========================================================================
   # Generate an example data model with a small molecule in it or else read
@@ -1595,15 +1615,10 @@ END
   doneInt = time.clock()
 
   print('Constructing Optimizer')
-  # @todo Let the caller specify the model index and altID rather than doing only the default (first).
   startOpt = time.clock()
-  opt = FastOptimizer(model,probeRadius=0.25, altID="")
+  opt = FastOptimizer(True, model,probeRadius=0.25)
   doneOpt = time.clock()
   info = opt.getInfo()
-  print('XXX info:\n'+info)
-  print('XXX Time to Add Hydrogen =',doneAdd-startAdd)
-  print('XXX Time to Interpret =',doneInt-startInt)
-  print('XXX Time to Optimize =',doneOpt-startOpt)
 
   f = open("deleteme.pdb","w")
   f.write(model.model_as_pdb())