+++ simtbx/diffBragg/tests/tst_diffBragg_utils.py
@@ -6,8 +6,9 @@ from simtbx.diffBragg.phil import phil_scope
 PARAMS = phil_scope.extract()
 
 import logging
+import sys
 logger = logging.getLogger()
-handler = logging.StreamHandler()
+handler = logging.StreamHandler(sys.stdout)
 format = "%(filename)s:%(funcName)s >> %(message)s"
 F = logging.Formatter(format)
 handler.setFormatter(fmt=F)